# controllers/Photo.js
'use strict';

var utils = require('../utils/writer.js');
var Photo = require('../service/PhotoService');

module.exports.photoGET = function photoGET (req, res, next) {
  var limit = req.swagger.params['limit'].value;
  var offset = req.swagger.params['offset'].value;
  var search = req.swagger.params['search'].value;
  Photo.photoGET(limit,offset,search)
    .then(function (response) {
      utils.writeJson(res, response);
    })
    .catch(function (response) {
      utils.writeJson(res, response);
    });
};